+++ CHANGELOG.md
@@ -24,6 +24,7 @@ TBD
 - Updated internal class from EventProcessor to ReplicationProcessor for improved clarity (#1780)
 - Updated latest sytem version to prioritize running but still return a system if not (#1949)
 - Updated Jobs pages to trigger off internal Job events for status changes (#1878, #1951)
+- Added metadata to Requests if they are Orphaned (#)
 
 
 ## 3.30.0rc2